Moved helpers class outside soh handler
