+++ custom/ewsghana/handlers/soh.py
@@ -1,10 +1,15 @@
-from collections import defaultdict, OrderedDict
+from collections import defaultdict
 
+from django.conf import settings
+
+from corehq.apps.commtrack.exceptions import NotAUserClassError, NoDefaultLocationException
 from corehq.apps.commtrack.models import StockState
+from corehq.apps.commtrack.sms import SMSError, process
+from corehq.apps.domain.models import Domain
 from corehq.apps.locations.dbaccessors import get_all_users_by_location
 from corehq.apps.reminders.util import get_preferred_phone_number_for_recipient
 from corehq.apps.users.models import CommCareUser
-from custom.ewsghana.handlers import INVALID_MESSAGE, INVALID_PRODUCT_CODE, ASSISTANCE_MESSAGE,\
+from custom.ewsghana.handlers import INVALID_MESSAGE, ASSISTANCE_MESSAGE,\
     MS_STOCKOUT, MS_RESOLVED_STOCKOUTS, NO_SUPPLY_POINT_MESSAGE
 from casexml.apps.stock.const import SECTION_TYPE_STOCK
 from casexml.apps.stock.models import StockTransaction
@@ -12,181 +17,14 @@ from corehq.apps.locations.models import SQLLocation
 from corehq.apps.products.models import SQLProduct
 from corehq.toggles import EWS_INVALID_REPORT_RESPONSE
 from custom.ewsghana.handlers.keyword import KeywordHandler
+from custom.ewsghana.handlers.helpers.formatter import EWSFormatter
+from custom.ewsghana.handlers.helpers.stock_and_receipt_parser import EWSStockAndReceiptParser, \
+    ProductCodeException
 from custom.ewsghana.reminders import ERROR_MESSAGE
 from custom.ewsghana.utils import ProductsReportHelper, send_sms
 from custom.ewsghana.alerts.alerts import SOHAlerts
-from corehq.apps.commtrack.sms import *
 from custom.ilsgateway.tanzania.reminders import SOH_HELP_MESSAGE
 from dimagi.utils.couch.database import iter_docs
-
-
-class ProductCodeException(Exception):
-    pass
-
-
-class EWSFormatter(object):
-
-    REC_SEPARATOR = '-'
-
-    def _clean_string(self, string):
-        if not string:
-            return string
-        mylist = list(string)
-        newstring = string[0]
-        i = 1
-        while i < len(mylist) - 1:
-            if mylist[i] == ' ' and mylist[i - 1].isdigit() and mylist[i + 1].isdigit():
-                newstring += self.REC_SEPARATOR
-            else:
-                newstring = newstring + mylist[i]
-            i += 1
-        newstring = newstring + string[-1]
-        string = newstring
-
-        string = string.replace(' ', '')
-        separators = [',', '/', ';', '*', '+', '-']
-        for mark in separators:
-            string = string.replace(mark, self.REC_SEPARATOR)
-        junk = ['\'', '\"', '`', '(', ')']
-        for mark in junk:
-            string = string.replace(mark, '')
-        return string.lower()
-
-    def _getTokens(self, string):
-        mylist = list(string)
-        token = ''
-        i = 0
-        while i < len(mylist):
-            token = token + mylist[i]
-            if i + 1 == len(mylist):
-                # you've reached the end
-                yield token
-            elif (mylist[i].isdigit() and not mylist[i + 1].isdigit()
-                  or mylist[i].isalpha() and not mylist[i + 1].isalpha()
-                  or not mylist[i].isalnum() and mylist[i + 1].isalnum()):
-                yield token
-                token = ''
-            i += 1
-
-    def format(self, string):
-        """
-        Old parse method, used in Ghana for more 'interesting' parsing.
-        Moved from: https://github.com/dimagi/rapidsms-logistics/blob/7a1433abbda4ec27dc8f4c5da14c0f5689abd202/logistics/models.py#L1430
-        """
-        if not string:
-            return
-        match = re.search("[0-9]", string)
-        if not match:
-            raise SMSError
-        string = self._clean_string(string)
-        an_iter = self._getTokens(string)
-        commodity = None
-        valid = False
-
-        product_quantity = OrderedDict()
-        while True:
-            try:
-                while commodity is None or not commodity.isalpha():
-                    commodity = an_iter.next().lower()
-                count = an_iter.next()
-                while not count.isdigit():
-                    count = an_iter.next()
-                product_quantity[commodity] = {'soh': count, 'receipts': 0}
-                valid = True
-                token_a = an_iter.next()
-                if not token_a.isalnum():
-                    token_b = an_iter.next()
-                    while not token_b.isalnum():
-                        token_b = an_iter.next()
-                    if token_b.isdigit():
-                        # if digit, then the user is reporting receipts
-                        product_quantity[commodity]['receipts'] = token_b
-                        commodity = None
-                        valid = True
-                    else:
-                        # if alpha, user is reporting soh, so loop
-                        commodity = token_b
-                        valid = True
-                else:
-                    commodity = token_a
-                    valid = True
-            except ValueError:
-                commodity = None
-                continue
-            except StopIteration:
-                break
-        if not valid:
-            return string
-
-        result = ""
-        for product, soh_receipt_dict in product_quantity.iteritems():
-            soh = soh_receipt_dict.get('soh')
-            if not soh:
-                continue
-            receipts = soh_receipt_dict.get('receipts', 0)
-            result += "{} {}.{} ".format(product, soh, receipts)
-
-        return result.strip()
-
-
-class EWSStockAndReceiptParser(StockAndReceiptParser):
-
-    def __init__(self, domain, v):
-        super(EWSStockAndReceiptParser, self).__init__(domain, v)
-        self.bad_codes = set()
-
-    def product_from_code(self, prod_code):
-        try:
-            return super(EWSStockAndReceiptParser, self).product_from_code(prod_code)
-        except SMSError:
-            return None
-
-    def single_action_transactions(self, action, args):
-        products = []
-        for idx, arg in enumerate(args):
-            if self.looks_like_prod_code(arg):
-                product = self.product_from_code(arg)
-                if product:
-                    products.append(product)
-                else:
-                    if idx == 0:
-                        raise ProductCodeException(INVALID_PRODUCT_CODE % arg)
-                    self.bad_codes.add(arg)
-            else:
-                if not products:
-                    continue
-                if len(products) > 1:
-                    raise SMSError('missing quantity for product "%s"' % products[-1].code)
-
-                # NOTE also custom code here, must be formatted like 11.22
-                if re.compile("^\d+\.\d+$").match(arg):
-                    value = arg
-                else:
-                    raise SMSError('could not understand product quantity "%s"' % arg)
-
-                for p in products:
-                    # for EWS we have to do two transactions, one being a receipt
-                    # and second being a transaction (that's reverse of the order
-                    # the user provides them)
-                    yield StockTransactionHelper(
-                        domain=self.domain.name,
-                        location_id=self.location.location_id,
-                        case_id=self.case_id,
-                        product_id=p.get_id,
-                        action=const.StockActions.RECEIPTS,
-                        quantity=Decimal(value.split('.')[1])
-                    )
-                    yield StockTransactionHelper(
-                        domain=self.domain.name,
-                        location_id=self.location.location_id,
-                        case_id=self.case_id,
-                        product_id=p.get_id,
-                        action=const.StockActions.STOCKONHAND,
-                        quantity=Decimal(value.split('.')[0])
-                    )
-                products = []
-        if products:
-            raise SMSError('missing quantity for product "%s"' % products[-1].code)
 
 
 def get_transactions_by_product(transactions):
